requirements.txt
crispy-bootstrap5==0.6
Django==4.0.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-leaflet==0.28.3
django-matplotlib==0.1
django-phonenumber-field==6.2.0
django-progressive-web-app==0.1.1
djangorestframework==3.13.1
gunicorn==20.1.0
phonenumbers==8.12.52
psycopg2-binary==2.9.3
dj-database-url==1.0.0
whitenoise==6.2.0
django-anymail==13.0
pandas
openpyxl
python-decouple==3.8
